+++ benchmarks/rendering/vite.config.ts
@@ -13,6 +13,7 @@ export default defineConfig({
     },
   },
   define: {
+    __GLUON_DEV__: JSON.stringify(false),
     __VUE_OPTIONS_API__: true,
     __VUE_PROD_DEVTOOLS__: false,
     __VUE_PROD_HYDRATION_MISMATCH_DETAILS__: false,